fix helperFunction
